Add video support from Twitter embeds +  Remove `u-photo` from Twitter photos
As found with jvt.me#883, a Tweet from Tech Nottingham included a GIF,
which under the `video` portion of the MF2 JSON, not the

An assumption we can make here is that the MIME type is going to be an
MP4 for everything, but that isn't ideal.

GIFs make sense to loop, and we want to limit the size of the `video`
element on the page, especially as they're not all built for size.

Closes #883.