Remove incorrect use of `p-name` on non-articles
As raised by aaronpk on the IndieWeb-dev chat, using the `p-name` here
is not correct, as it should be only used for the title of a post. In
these cases, these content types do not have a title (as that is
reserved for articles), just content.

This will screw with post-type-discovery, as that will return `article`
when it finds a `p-name`.

In the case of RSVPs, we can remove the `p-name` completely, as it was
meant to be hidden metadata, which now no longer makes sense.

Closes #894.