Fix: Store meetup-mf2 event locations correctly
I didn't notice until the other day that RSVPs via meetup-mf2 were
rendering with an invalid location.

It appears that this is due to the response from meetup-mf2 coming back
correctly as a `String[]`, but being put into my site as a `String`.

This was originally picked up as part of meetup-mf2!5 but wasn't fully
actioned until www-api!104.

The fix in the site is to make sure that we store each location as a
`String[]`.

Closes #925.