Fix: Align Twitter -contexts better
When using `inline-block`, this breaks the `single.html` rendering, as
the Twitter context displays 'inline' with the left side of the browser.

By replacing it with the default, which is `display: block`, it'll
render in the middle of the container on `list.html`s, and correctly on
the `single.html`s.

Closes #935.