Add note on /404/ to warn usage of illegal querystring
I've had a number of hits from folks using invalid URLs, so it'd be nice
to at least have something warning them why it's broken.

We can detect this by checking if there's nothing that can be parsed
from the querystring (via `URLSearchParams`) but that there is a `&` in
the URL, which will be either a querystring or a URL fragment.

Closes #853.