Use `dind` service on self-hosted master builds
To see if I can speed up deploys, I want to use the self-hosted GitLab
Runner. However, there is still an issue around not reusing the
cache[0].

It appears a solution to the issue around caching concerns is to
have a long-lived `dind` service that runs on the host, and will cache
images, which I've creatively named `dind`.

[0]: https://gitlab.com/gitlab-org/gitlab-foss/issues/17861.