Render u-syndication links for /mf2/ better
The u-syndication links are meant to point to the syndicated post, not
the place that should syndicate it to (although it's largely because
we're abusing Webmentions for syndication purposes).

- For posts with actual permalinks, such as
  Twitter/Meetup/GitHub/IndieNews, we link to the permalink
- For Bridgy links, we hide them, so they're present to send
  Webmentions, but don't show anything to the user
- If I don't have a fancy way to display the link, render the hostname

We also want to remove the `<details>` tag, as these should be visible
to folks who want to see the posts elsewhere.

This is easier to manage for anything on /mf2/ so we'll just work with
that for now.