Align rendering of the feed across kinds
To make it a bit nicer between articles and /mf2/ entries, we should
move to using a more consistent feel between kinds.

Firstly this requires making the presentation of the articles smaller,
and a bit less bold in the feed, while taking care to preserve the
Microformats2 content of the post.

Articles' titles should be a bit bolder than regular text, but not as
big as an `h1` previously.

Following influence from https://mrkapowski.com/ we can decorate the
post with details of the kind of post it is.  To make it more
extensible, we can create a new `kinds` data setup, which can contain
information about each of the kinds, such as the icon.