Render the -context for posts as an h-cite
I added the -context for Tweets to the rendering on my site so viewers
would be able to at least have some idea what was being interacted with,
instead of having to i.e. follow the link and see for themselves.

However, this wasn't represented for folks following my site in a
Microformats2 feed. This data is now exposed with Microformats2 as an
`h-cite`.

This requires a few things:

- We need to make sure that the context is rendered with Microformats2
  markup inside
- We need to make sure that i.e. the `u-like-of` is the same container
  as the `h-cite`

Closes #949.