Add support for firehose feed
To reduce the amount of noise on my main feed, that most folks aren't
interested in, we can create a separate feed.

On the firehose feed, nothing is ignored, whereas on the core feed, we
only want to render certain posts.

This requires adding a new link to the menu, as well as to the main
feed.