Add public listing of contacts for `q=contact`
As part of www-api#105, I want to implement contact management for my
site without having to manually push the config myself.

As well as being able to add contacts, I also want to allow other
clients to consume it, so need to generate this static content that can
then be proxied via Micropub.

Because we want to have a URL that can be utilised by Micropub for
updates, and may as well have it as a real page on the site. So we need
to migrate the existing content in a data template to a regular section,

While migrating them, we should migrate them to a format that is
allowable under a `HugoMicroformats2` representation, so it's easier to
work with in Micropub, as well as if we want to migrate them to a
different site in the future.

However, we will not implement the actual pages for now, leaving it to #960.