Render h-card image side-by-side
It's much nicer to view the image side-by-side (which is super easy
using flexbox) rather than inline.

This currently looks quite nice with the amount of text that is shown,
and if there's more text, there's just a bit of blank space under the
text.

This looks alright on mobile, although could be improved in the future.