Remove GitLab CI deployments to Netlify
I'm going to try using Netlify-only deployments for the site, to see if
the builds are any more reliable / quicker.

As such, we may as well remove any GitLab-specific configuration, as it
doesn't need to be here any more.

We can also remove our Git submodule for the helper tooling for
deploying to Spectat infrastructure.