Fix: Correctly nest `h-cite`s
As noticed in #989, the rendering of Twitter -contexts has been broken
when there are nested `h-cite`s, as the child element does not have a
separate `h-cite`, meaning a Microformats2 parser (i.e. for Bridgy) will
pick up the child post, too, which leads to me to posting correctly to
my site, but syndication interacts with the wrong content - which is
bad!