Render event permalink if one exists
When I've created a self-owned event, I want to be able to specify a
custom URL, i.e. to point it towards the real URL, but that doesn't get
populated in the `in-reply-to` as we never provide the `u-url` of the
event in the markup.

By adding this, RSVPs will now be marked up more correctly.