Move `music-in-review` to its own section
Similar to the work we've done with week notes, these can be stored
separately, which makes it handier as it doesn't bleed into the
directory structure for "regular" articles.

We can replace the existing URL structure, while preserving permalinks,
with a more applicable structure.

We can also replace a number of repeated pieces of metadata with
defaulting it and then `cascade`ing it.

This also requires a theme change to ignore the `_index.md` that's
added, otherwise it would be rendered in taxonomy pages.