Add ability to deploy to S3 using `hugo deploy`
As a way to investigate whether my site would be better served from
S3 + CloudFront, deployment speed wise, I want to see what happens when
each and every deploy goes through S3, and what the speed implications
are.

This utilises Hugo's inbuilt `deploy` functionality, which is pretty
damn cool, and doesn't need any of the AWS CLI/SDK available, as it's
bundled in Hugo (extended).

We need to make sure we use `hugo-official` for the deploy, as the base
image provides a wrapper for `hugo` that sets other options, which we
can't use for the `deploy` subcommand.