Reduce size of produced HTML files
To simplify uploads, and speed up shipping time, we want to limit the
size of any HTML files in the site.

This also allows us to configure the maximum number of items in an RSS
feed, for both size reduction, and to prevent readers' readers not being
happy with the size of them.