Soft redirect deleted pages to /gone.html
As noted in #1055, we're seeing deleted posts in Production.

This is because Netlify used to manage this, but S3+CloudFront doesn't
have any of these custom redirects available.

What we can do, especially for local testing, is to use a soft-redirect
for the deleted pages to `/gone.html`.

There will be some addition code at the CloudFront layer which will see
this and correctly return HTTP 410 Gone.

Closes #1055.