Don't allow concurrent deploys
To prevent concurrent deploys messing up the S3 bucket - especially when
an older build runs after a later build has - we can use
`resource_group`s to force a lock on the Prod deploy.

This requires separate stages, so we ned to add the `public` folder as
an artifact so it can be re-used on a later build.

Because they're only used for the next stage, they don't need to live
any longer than 30 minutes.