Re-create popular-posts page
It appears that we can't use `$.Site.GetPage` from a partial via the
`partial` shortcode, so we can add a new shortcode instead.

Note that this isn't quite "re-creating" it, as this is a different
intent of the previous one, where this now is curated, instead of
automagic, but this is good for now.